fix: set base URI on the parser which reads the API docs graph
