+++ api/index.ts
@@ -1,15 +1,14 @@
 import cors from 'cors'
 import path from 'path'
-import hydraBox from 'hydra-box'
 import url from 'url'
 import express from 'express'
-import replace from 'replace-in-file'
 import dotenv from 'dotenv'
 import dotenvExpand from 'dotenv-expand'
 import debug from 'debug'
 import { NotFoundError } from './lib/error'
 import { httpProblemMiddleware } from './lib/error/middleware'
 import frontend, { rootRedirect } from './frontend'
+import hydraMiddleware from './lib/hydra-box'
 
 dotenvExpand(dotenv.config())
 import('./lib/handlers')
@@ -33,33 +32,6 @@ function logger (req: express.Request, res, next) {
   next()
 }
 
-async function hydraMiddleware () {
-  let authentication
-  if (process.env.SPARQL_ENDPOINT_USERNAME && process.env.SPARQL_ENDPOINT_PASSWORD) {
-    authentication = {
-      user: process.env.SPARQL_ENDPOINT_USERNAME,
-      password: process.env.SPARQL_ENDPOINT_PASSWORD,
-    }
-  }
-
-  const options: Record<string, unknown> = {
-    debug: true,
-    sparqlEndpointUrl: process.env.READ_MODEL_SPARQL_ENDPOINT,
-    sparqlEndpointUpdateUrl: process.env.SPARQL_UPDATE_ENDPOINT,
-    contextHeader: '/context/',
-    authentication,
-  }
-
-  const apiDocsPath = path.join(__dirname, 'hydra/api.ttl')
-  await replace({
-    files: apiDocsPath,
-    from: /@base <.+>/,
-    to: `@base <${process.env.BASE_URI}>`,
-  })
-
-  return hydraBox.fromUrl('/api', 'file://' + apiDocsPath, options)
-}
-
 Promise.resolve().then(async () => {
   const baseUrl = `${process.env.BASE_URI}`
 
@@ -74,7 +46,7 @@ Promise.resolve().then(async () => {
   app.use(cors({
     exposedHeaders: ['link', 'location'],
   }))
-  app.use(await hydraMiddleware())
+  app.use(await hydraMiddleware(path.join(__dirname, 'hydra/api.ttl')))
   app.use(function (req, res, next) {
     next(new NotFoundError())
   })
